api/requirements.txt
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4 